# openapi.yaml
...
openapi: "3.1.2"
info:
  title: Apache Nutch REST API
  description: >-
    REST API for managing Apache Nutch crawl jobs, configurations, seed lists,
    database queries, and data readers. The server is powered by Apache CXF
    with JAX-RS.
  version: "1.0.0"
  license:
    name: Apache 2.0
    identifier: Apache-2.0
  contact:
    name: Apache Nutch
    url: https://nutch.apache.org
servers:
  - url: "{protocol}://localhost:{port}"
    description: Nutch REST server
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
        description: The protocol used to access the Nutch server.
      port:
        default: "8081"
        description: >-
          The port the Nutch server listens on. Configurable via the --port
          command-line argument.
security:
  - basicAuth: []
...
paths:
...
  /config/{configId}/{propertyId}:
    get:
      tags:
        - Configuration
      summary: Get a single configuration property
      description: >-
        Returns the value of a specific property within the given
        configuration.
      operationId: getProperty
      parameters:
        - $ref: "#/components/parameters/configId"
        - $ref: "#/components/parameters/propertyId"
      responses:
        "200":
          description: The property value as plain text.
          content:
            text/plain:
              schema:
                type: string
              example: "NutchBot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Configuration
      summary: Update a configuration property
      description: >-
        Adds or updates the value of a property in the specified
        configuration.
      operationId: updateProperty
      parameters:
        - $ref: "#/components/parameters/configId"
        - $ref: "#/components/parameters/propertyId"
      requestBody:
        required: true
        description: The new property value as plain text.
        content:
          text/plain:
            schema:
              type: string
            example: "MyNewBot"
      responses:
        "200":
          description: Property updated successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /seed/:
    get:
      tags:
        - Seed
      summary: List all seed lists
      description: Returns a map of all created seed files keyed by name.
      operationId: getSeedLists
      responses:
        "200":
          description: A JSON object mapping seed list names to SeedList objects.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/SeedList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /seed/create:
    post:
      tags:
        - Seed
      summary: Create a seed list file
      description: >-
        Creates a seed list file from the provided URLs and writes it to
        HDFS. Returns the path to the created seed file directory.
      operationId: createSeedFile
      requestBody:
        required: true
        description: The seed list containing URLs to write.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedList"
      responses:
        "200":
          description: Path to the created seed file directory.
          content:
            text/plain:
              schema:
                type: string
              example: "seedFiles/seed-1700000000000"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Reader
  # ---------------------------------------------------------------------------
  /reader/sequence/read:
    post:
      tags:
        - Reader
      summary: Read a sequence file
      description: >-
        Reads key-value pairs from a Hadoop sequence file. Supports reading
        all rows, a limited number of rows, a row range, or counting the
        total number of rows.
      operationId: seqRead
      parameters:
        - $ref: "#/components/parameters/nrows"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/count"
      requestBody:
        required: true
        description: Reader configuration specifying the file path.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderConfig"
      responses:
        "200":
          description: >-
            Sequence file data. Returns application/json when reading rows,
            or text/plain when count=true.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
            text/plain:
              schema:
                type: integer
                format: int32
                minimum: 0
                maximum: 2147483647
                description: Number of rows in the sequence file.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /reader/link:
    get:
      tags:
        - Reader
      summary: Get link reader schema
      description: >-
        Returns the schema describing the fields in link reader responses.
      operationId: getLinkSchema
      responses:
        "200":
          description: Link reader response schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkSchema"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /reader/link/read:
    post:
      tags:
        - Reader
      summary: Read link objects
      description: >-
        Reads link data (LinkDatum) from the Nutch webgraph. Supports
        reading all rows, a limited number of rows, a row range, or
        counting the total number of rows.
      operationId: linkRead
      parameters:
        - $ref: "#/components/parameters/nrows"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/count"
      requestBody:
        required: true
        description: Reader configuration specifying the file path.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderConfig"
      responses:
        "200":
          description: >-
            Link data. Returns application/json when reading rows, or
            text/plain when count=true.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
            text/plain:
              schema:
                type: integer
                format: int32
                minimum: 0
                maximum: 2147483647
                description: Number of link entries.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /reader/node:
    get:
      tags:
        - Reader
      summary: Get node reader schema
      description: >-
        Returns the schema describing the fields in node reader responses.
      operationId: getNodeSchema
      responses:
        "200":
          description: Node reader response schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeSchema"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /reader/node/read:
    post:
      tags:
        - Reader
      summary: Read node objects
      description: >-
        Reads Node objects from the Nutch webgraph. Supports reading all
        rows, a limited number of rows, a row range, or counting the total
        number of rows.
      operationId: nodeRead
      parameters:
        - $ref: "#/components/parameters/nrows"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/count"
      requestBody:
        required: true
        description: Reader configuration specifying the file path.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderConfig"
      responses:
        "200":
          description: >-
            Node data. Returns application/json when reading rows, or
            text/plain when count=true.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
            text/plain:
              schema:
                type: integer
                format: int32
                minimum: 0
                maximum: 2147483647
                description: Number of node entries.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Services
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    configId:
      name: configId
      in: path
      required: true
      description: The unique identifier for the configuration.
      schema:
        type: string
    propertyId:
      name: propertyId
      in: path
      required: true
      description: The name (key) of the configuration property.
      schema:
        type: string
...
    nrows:
      name: nrows
      in: query
      required: false
      description: >-
        Number of rows to read. If not specified (or -1), all rows are
        returned.
      schema:
        type: integer
        format: int32
        minimum: -1
        maximum: 2147483647
        default: -1
    start:
      name: start
      in: query
      required: false
      description: Starting line number for a range read.
      schema:
        type: integer
        format: int32
        minimum: -1
        maximum: 2147483647
        default: -1
    end:
      name: end
      in: query
      required: false
      description: Ending line number for a range read.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 2147483647
    count:
      name: count
      in: query
      required: false
      description: >-
        If true, returns the number of lines instead of the data itself.
        When set, the response content type is text/plain.
      schema:
        type: boolean
        default: false

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ReaderConfig:
      type: object
      description: Configuration specifying a file path for reader operations.
      required:
        - path
      properties:
        path:
          type: string
          description: >-
            The path to the sequence file, link data, or node data to read.
      example:
        path: "crawl-01/crawldb/current/part-00000/data"
    SeedList:
      type: object
      description: A named list of seed URLs.
      required:
        - seedUrls
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          maximum: 9007199254740991
          description: The seed list identifier.
          readOnly: true
        name:
          type: string
          description: A human-readable name for this seed list.
        seedFilePath:
          type: string
          description: >-
            The HDFS path where the seed file is stored. Populated after
            creation.
          readOnly: true
        seedUrls:
          type: array
          items:
            $ref: "#/components/schemas/SeedUrl"
          description: The collection of seed URLs in this list.
      example:
        name: "my-seeds"
        seedUrls:
          - url: "https://example.com"
          - url: "https://nutch.apache.org"
    SeedUrl:
      type: object
      description: A single seed URL entry.
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          maximum: 9007199254740991
          description: The seed URL identifier.
          readOnly: true
        url:
          type: string
          description: The seed URL.
      example:
        url: "https://example.com"
...
    LinkSchema:
      type: object
      description: Schema describing the fields in a link reader response.
      properties:
        key_url:
          type: string
          example: "string"
        timestamp:
          type: string
          example: "int"
        score:
          type: string
          example: "float"
        anchor:
          type: string
          example: "string"
        linktype:
          type: string
          example: "string"
        url:
          type: string
          example: "string"
    NodeSchema:
      type: object
      description: Schema describing the fields in a node reader response.
      properties:
        key_url:
          type: string
          example: "string"
        num_inlinks:
          type: string
          example: "int"
        num_outlinks:
          type: string
          example: "int"
        inlink_score:
          type: string
          example: "float"
        outlink_score:
          type: string
          example: "float"
        metadata:
          type: string
          example: "string"

    # -- Enums ----------------------------------------------------------------
...
  responses:
    BadRequest:
      description: >-
        Bad request. The request body is missing, malformed, or contains
        invalid parameters.
      content:
        text/plain:
          schema:
            type: string
          example: "Nutch configuration cannot be empty!"
    Unauthorized:
      description: >-
        Unauthorized. Basic authentication credentials are missing or
        invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Authentication required."
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Resource not found."
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        text/plain:
          schema:
            type: string
          example: "Internal server error."